Move SMMSTOREv1 to a module
There are 2 implementations of SMMSTORE: v1 and v2. Move the existing
v1-only logic to a module in preparation of adding v2 logic.
